chore: Move UI tests to separate package
This allows us to lower the MSRV
