--- yare-macro/tests/fail/invalid_group_brackets.rs
@@ -1,8 +1,0 @@
-use yare_macro::parameterized;
-
-#[parameterized(
-    case1 = [ 0 ],
-)]
-fn test(some: u32, more: u32) {}
-
-fn main() {}
