--- yare-macro/tests/ok/qualifier_const.rs
@@ -1,8 +1,0 @@
-use yare_macro::parameterized;
-
-#[parameterized(
-    case1 = { 0 },
-)]
-const fn test(a: u32) {}
-
-fn main() {}
