+++ yare-tests-ui/tests/fail/invalid_group_brackets.rs
@@ -1,7 +1,7 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
-    case1 = ( 0 ),
+    case1 = [ 0 ],
 )]
 fn test(some: u32, more: u32) {}
