+++ yare-tests-ui/tests/fail/invalid_group_brackets.stderr
