+++ yare-tests-ui/tests/fail/invalid_group_parenthesises.rs
@@ -1,7 +1,7 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
-    case1 = { 0, 1, 2 },
+    case1 = ( 0 ),
 )]
 fn test(some: u32, more: u32) {}
