+++ yare-tests-ui/tests/fail/not_a_fn.rs
@@ -1,8 +1,8 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = { 0 },
 )]
-fn test(a: u32) {}
+struct A;
 
 fn main() {}
