+++ yare-tests-ui/tests/fail/not_enough_arguments_2nd.rs
@@ -1,8 +1,8 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = { 0, 1 },
-    case2 = { 0, 1 },
+    case2 = { 0 },
 )]
 fn test(some: u32, more: u32) {}
