+++ yare-tests-ui/tests/fail/not_enough_arguments_multi_case.rs
@@ -1,4 +1,4 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = { 0 },
