+++ yare-tests-ui/tests/fail/not_enough_arguments_multi_case.stderr
