+++ yare-tests-ui/tests/fail/too_many_arguments.rs
@@ -1,8 +1,7 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = { 0, 1, 2 },
-    case2 = { 0, 1, 2 },
 )]
 fn test(some: u32, more: u32) {}
