+++ yare-tests-ui/tests/fail/too_many_arguments_2nd.rs
@@ -1,0 +1,9 @@
+use yare::parameterized;
+
+#[parameterized(
+    case = { 0, 1 },
+    case2 = { 0, 1, 2 },
+)]
+fn test(some: u32, more: u32) {}
+
+fn main() {}
