+++ yare-tests-ui/tests/fail/too_many_arguments_multi_case.rs
@@ -1,7 +1,7 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
-    case = { 0, 1 },
+    case1 = { 0, 1, 2 },
     case2 = { 0, 1, 2 },
 )]
 fn test(some: u32, more: u32) {}
