+++ yare-tests-ui/tests/ok/multiple_parameterized_tests.rs
@@ -1,4 +1,4 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = {},
