+++ yare-tests-ui/tests/ok/one_param.rs
@@ -1,0 +1,8 @@
+use yare::parameterized;
+
+#[parameterized(
+    case1 = { 0 },
+)]
+fn test(a: u32) {}
+
+fn main() {}
