+++ yare-tests-ui/tests/ok/should_panic.rs
@@ -1,4 +1,4 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = {},
