+++ yare-tests-ui/tests/ok/trailing_commas_between_arguments.rs
@@ -1,4 +1,4 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = { 0, 1, },
