+++ yare-tests-ui/tests/ok/trailing_commas_between_cases.rs
@@ -1,8 +1,8 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = { 0, 1 },
-    case2 = { 0 },
+    case2 = { 0, 1 },
 )]
 fn test(some: u32, more: u32) {}
