+++ yare-tests-ui/tests/ok/two_params.rs
@@ -1,4 +1,4 @@
-use yare_macro::parameterized;
+use yare::parameterized;
 
 #[parameterized(
     case1 = { 0, 0 },
